# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Glucometer Tracker API",
    "description": "API documentation for the Glucometer Tracker project with session-based authentication and GitHub OAuth.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/users/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Create an account using username, email, and password. Passwords are hashed with bcrypt before storage.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string", "example": "newuser" },
                "email": { "type": "string", "example": "user@example.com" },
                "password": { "type": "string", "example": "SecurePass123" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Validation failed or user already exists" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Login a user",
        "description": "Authenticate using email and password and establish a session cookie.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string", "example": "user@example.com" },
                "password": { "type": "string", "example": "SecurePass123" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/users/logout": {
      "get": {
        "summary": "Logout current user",
        "description": "Destroy the current session and log the user out.",
        "responses": {
          "200": { "description": "Logout successful" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/auth/github": {
      "get": {
        "summary": "Login with GitHub",
        "description": "Redirects the browser to GitHub for OAuth login. This endpoint cannot be successfully executed via Swagger UI's Try It Out because it performs a cross-origin redirect.",
        "responses": {
          "302": { "description": "Redirect to GitHub OAuth" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/auth/github/callback": {
      "get": {
        "summary": "GitHub OAuth callback",
        "description": "GitHub redirects back to this URL after authentication. The callback will establish a server session and redirect to the application root.",
        "responses": {
          "302": { "description": "Redirect to application root after successful login" },
          "401": { "description": "OAuth authentication failed" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/status": {
      "get": {
        "summary": "Check authentication status",
        "description": "Returns whether the client is currently logged in.",
        "responses": {
          "200": { "description": "Status returned" }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "description": "Protected endpoint, requires an active session.",
        "security": [ { "sessionAuth": [] } ],
        "responses": {
          "200": { "description": "List of users" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get a user by ID",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "User found" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "User deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/users/{id}/settings": {
      "put": {
        "summary": "Update user settings",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "settings": {
                  "type": "object",
                  "properties": {
                    "lowThreshold": { "type": "number" },
                    "highThreshold": { "type": "number" },
                    "units": { "type": "string", "enum": ["mg/dL", "mmol/L"] }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Settings updated" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/readings": {
      "get": {
        "summary": "Get all readings",
        "description": "Protected endpoint requiring an authenticated session.",
        "security": [ { "sessionAuth": [] } ],
        "responses": {
          "200": { "description": "List of readings" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "summary": "Create a new reading",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": { "type": "string", "example": "60fa0cdd4af8e411c8a2b1c1" },
                "glucoseLevel": { "type": "number", "example": 5.4 },
                "context": {
                  "type": "object",
                  "properties": {
                    "meal": { "type": "string", "example": "Breakfast" },
                    "exercise": { "type": "string", "example": "30 min walk" },
                    "notes": { "type": "string", "example": "Felt good" }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Reading created" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/readings/user/{userId}": {
      "get": {
        "summary": "Get readings for a user",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "User readings" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/readings/{id}": {
      "get": {
        "summary": "Get a reading by ID",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Reading details" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "summary": "Update a reading",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "glucoseLevel": { "type": "number" },
                "context": {
                  "type": "object",
                  "properties": {
                    "meal": { "type": "string" },
                    "exercise": { "type": "string" },
                    "notes": { "type": "string" }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Reading updated" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "summary": "Delete a reading",
        "security": [ { "sessionAuth": [] } ],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Reading deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  },
  "securityDefinitions": {
    "sessionAuth": {
      "type": "apiKey",
      "name": "connect.sid",
      "in": "cookie",
      "description": "Session cookie used for authenticated requests after login."
    }
  }
}